suggested code review changes from GY (combine path_functions & tag_utility -> dls_utilities; tag_check function returns a bool)
