+++ dls_ade/dls_changes_since_release.py
@@ -7,7 +7,7 @@ import shutil
 import logging
 
 from dls_ade.argument_parser import ArgParser
-from dls_ade import path_functions as pathf
+from dls_ade import dls_utilities as pathf
 from dls_ade import vcs_git
 from dls_ade import Server
 from dls_ade import logconfig
