+++ dls_ade/dls_tar_module.py
@@ -18,7 +18,7 @@ from dls_ade import vcs_git
 from dls_ade.dls_environment import environment
 from dls_ade.argument_parser import ArgParser
 from dls_ade import dlsbuild
-from dls_ade import path_functions as pathf
+from dls_ade import dls_utilities as pathf
 from dls_ade import logconfig
 
 env = environment()
