+++ dls_ade/dls_utilities.py
@@ -1,4 +1,4 @@
-import os
+import os, re
 from dls_ade.exceptions import ParsingError
 
 GIT_ROOT_DIR = os.getenv('GIT_ROOT_DIR', "controls")
@@ -28,3 +28,24 @@ def check_technical_area(area, module):
 
     if area == "ioc" and len(module.split('/')) < 2:
         raise ParsingError("Missing technical area under beamline")
+
+
+def check_tag_is_valid(tag):
+    """
+    Checks if a given tag is a valid tag.
+
+    Args:
+        tag(str): proposed tag string
+
+    Returns:
+        bool: True if tag is valid, False if not
+
+    """
+
+    check = re.compile('[0-9]+\-[0-9]+(\-[0-9]+)?(dls[0-9]+(\-[0-9]+)?)?')
+    result = check.match(tag)
+
+    if result is None or result.group() != tag:
+        return False
+
+    return True
