+++ dls_ade/dls_utilities_test.py
@@ -1,28 +1,45 @@
-from dls_ade import path_functions
+from dls_ade import dls_utilities
 import unittest
+
 from dls_ade.exceptions import ParsingError
 
+from dls_ade.dls_utilities import check_tag_is_valid
+
+
+class TagFormatTest(unittest.TestCase):
+
+    def test_approves_correct_tags(self):
+
+        ex_valid_tags = ['2-51', '21-5-3', '2-5dls1', '2-5dls2-3']
+        for ex_valid_tag in ex_valid_tags:
+            self.assertTrue(check_tag_is_valid(ex_valid_tag))
+
+    def test_rejects_incorrect_tags(self):
+
+        ex_invalid_tags = ['2-5a', 'abcdef']
+        for ex_invalid_tag in ex_invalid_tags:
+            self.assertFalse(check_tag_is_valid(ex_invalid_tag))
 
 class RemoveEndSlash(unittest.TestCase):
 
     def test_given_empty_string_then_return(self):
         test_string = ""
 
-        new_string = path_functions.remove_end_slash(test_string)
+        new_string = dls_utilities.remove_end_slash(test_string)
 
         self.assertEqual(new_string, "")
 
     def test_given_path_slash_then_removed_and_returned(self):
         path = "controls/area/module/"
 
-        new_path = path_functions.remove_end_slash(path)
+        new_path = dls_utilities.remove_end_slash(path)
 
         self.assertEqual(new_path, "controls/area/module")
 
     def test_given_path_no_slash_then_returned(self):
         path = "controls/area/module"
 
-        new_path = path_functions.remove_end_slash(path)
+        new_path = dls_utilities.remove_end_slash(path)
 
         self.assertEqual(new_path, path)
 
@@ -33,13 +50,13 @@ class CheckTechnicalAreaValidTest(unittest.TestCase):
         area = "support"
         module = "test_module"
 
-        path_functions.check_technical_area(area, module)
+        dls_utilities.check_technical_area(area, module)
 
     def test_given_area_ioc_module_split_two_then_no_error_raised(self):
         area = "ioc"
         module = "modules/test_module"
 
-        path_functions.check_technical_area(area, module)
+        dls_utilities.check_technical_area(area, module)
 
     def test_given_area_ioc_module_split_less_than_two_then_no_error_raised(self):
         area = "ioc"
@@ -47,6 +64,6 @@ class CheckTechnicalAreaValidTest(unittest.TestCase):
         expected_error_msg = "Missing technical area under beamline"
 
         try:
-            path_functions.check_technical_area(area, module)
+            dls_utilities.check_technical_area(area, module)
         except ParsingError as error:
             self.assertEqual(str(error), expected_error_msg)
